chore: fix some imports
