+++ src/modules/users/infra/index.ts
@@ -1,3 +1,4 @@
 export * from "./controllers";
 export * from "./repositories";
 export * from "./routes";
+export * from "./factories";
