+++ src/modules/users/infra/repositories/index.ts
@@ -1,1 +1,2 @@
-export * from "./IUserRepository";
+export * from "./UserRepository";
+export * from "./UserRepository";
